move TeamEvalResult to datatransfer
TeamEvalResult does not communicate with any *Db and thus is wrong to be in core logic; in fact it has been functioning as a DTO from the beginning.
